pyproject.toml
[project]
name = "whatsapp-claude-bridge"
version = "0.1.0"
description = "Two-way WhatsApp chat with Claude AI including bash execution and conversation memory"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlmodel>=0.0.24",
    "twilio>=9.6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
